Bipartite random signed graphs

Bipartite random signed graphs

Usage

sample_bipartite_signed(
  n1,
  n2,
  p,
  p_neg,
  directed = FALSE,
  mode = c("out", "in", "all")
)

Arguments

n1

Integer scalar, the number of bottom vertices.

n2

Integer scalar, the number of top vertices.

p

The probability for drawing an edge between two arbitrary vertices.

p_neg

The probability of a drawn edge to be a negative tie

directed

logical, whether the graph will be directed. defaults to FALSE.

mode

Character scalar, specifies how to direct the edges in directed graphs. If it is ‘out’, then directed edges point from bottom vertices to top vertices. If it is ‘in’, edges point from top vertices to bottom vertices. ‘out’ and ‘in’ do not generate mutual edges. If this argument is ‘all’, then each edge direction is considered independently and mutual edges might be generated. This argument is ignored for undirected graphs.

Value

A signed bipartite igraph graph.

Examples

sample_bipartite_signed(10, 10, 0.5, 0.5)
#> IGRAPH a0d9e61 U--B 20 48 -- Bipartite Gnp random graph
#> + attr: name (g/c), p (g/n), type (v/l), sign (e/n)
#> + edges from a0d9e61:
#>  [1]  2--11  3--11  4--11  6--11  1--12  2--12  3--12  4--12 10--12  1--13
#> [11]  4--13  7--13  9--13 10--13  2--14  6--14  7--14  8--14 10--14  6--15
#> [21]  7--15  8--15  9--15  3--16  4--16  8--16  9--16  1--17  3--17  7--17
#> [31]  9--17  4--18  5--18 10--18  2--19  3--19  5--19  6--19  7--19  8--19
#> [41]  9--19 10--19  1--20  2--20  3--20  6--20  7--20  9--20