A graph with random subgraphs connected by negative edges

Create a number of Erdos-Renyi random graphs with identical parameters, and connect them with the specified number of negative ties.

Usage

sample_islands_signed(islands.n, islands.size, islands.pin, n.inter)

Arguments

islands.n

The number of islands in the graph.

islands.size

The size of the islands in the graph.

islands.pin

The probability of intra-island edges.

n.inter

number of negative edges between two islands.

Value

a signed igraph graph

Author

David Schoch

Examples

library(igraph)
sample_islands_signed(3, 10, 0.5, 1)
#> IGRAPH aaf744a U--- 30 70 -- 
#> + attr: grp (v/c), sign (e/n)
#> + edges from aaf744a:
#>  [1]  1-- 3  1-- 4  1-- 5  1-- 6  1-- 8  1--10  2-- 3  2-- 4  2-- 5  2-- 7
#> [11]  2-- 9  3-- 7  3-- 8  3-- 9  4-- 5  4-- 6  4-- 8  5-- 7  5-- 8  5-- 9
#> [21]  6-- 8  7-- 9  7--10 11--12 11--13 11--15 11--16 11--18 11--20 12--14
#> [31] 12--15 12--18 12--19 13--16 13--20 14--16 14--19 14--20 15--18 15--19
#> [41] 15--20 16--18 17--18 17--19 21--23 21--24 21--25 21--27 21--28 21--29
#> [51] 21--30 22--26 22--27 22--30 23--24 23--26 23--27 23--29 24--25 24--29
#> [61] 24--30 25--27 25--29 26--27 26--28 27--30 28--29  2--18 20--23  4--30